deploy: move cli-related packages in cmd
