+++ pkg/deploy/cmd/reaper_test.go
@@ -1,4 +1,4 @@
-package reaper
+package cmd
 
 import (
 	"reflect"
