+++ pkg/deploy/cmd/test/support.go
