--- pkg/deploy/reaper/doc.go
@@ -1,2 +1,0 @@
-// Package reaper implements the Reaper interface for deploymentConfigs
-package reaper
