--- pkg/deploy/scaler/doc.go
@@ -1,2 +1,0 @@
-// Package scaler implements the kubectl.Scaler interface for deploymentConfigs
-package scaler
