+++ pkg/deploy/strategy/recreate/recreate_test.go
@@ -12,7 +12,7 @@ import (
 	deployapi "github.com/openshift/origin/pkg/deploy/api"
 	deploytest "github.com/openshift/origin/pkg/deploy/api/test"
 	deployv1 "github.com/openshift/origin/pkg/deploy/api/v1"
-	scalertest "github.com/openshift/origin/pkg/deploy/scaler/test"
+	cmdtest "github.com/openshift/origin/pkg/deploy/cmd/test"
 	"github.com/openshift/origin/pkg/deploy/strategy"
 	deployutil "github.com/openshift/origin/pkg/deploy/util"
 
@@ -21,7 +21,7 @@ import (
 
 func TestRecreate_initialDeployment(t *testing.T) {
 	var deployment *kapi.ReplicationController
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 	strategy := &RecreateDeploymentStrategy{
 		out:          &bytes.Buffer{},
 		errOut:       &bytes.Buffer{},
@@ -55,7 +55,7 @@ func TestRecreate_deploymentPreHookSuccess(t *testing.T) {
 	config := deploytest.OkDeploymentConfig(1)
 	config.Spec.Strategy = recreateParams(30, deployapi.LifecycleHookFailurePolicyAbort, "", "")
 	deployment, _ := deployutil.MakeDeployment(config, kapi.Codecs.LegacyCodec(registered.GroupOrDie(kapi.GroupName).GroupVersions[0]))
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	hookExecuted := false
 	strategy := &RecreateDeploymentStrategy{
@@ -90,7 +90,7 @@ func TestRecreate_deploymentPreHookFail(t *testing.T) {
 	config := deploytest.OkDeploymentConfig(1)
 	config.Spec.Strategy = recreateParams(30, deployapi.LifecycleHookFailurePolicyAbort, "", "")
 	deployment, _ := deployutil.MakeDeployment(config, kapi.Codecs.LegacyCodec(registered.GroupOrDie(kapi.GroupName).GroupVersions[0]))
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	strategy := &RecreateDeploymentStrategy{
 		out:          &bytes.Buffer{},
@@ -123,7 +123,7 @@ func TestRecreate_deploymentMidHookSuccess(t *testing.T) {
 	config := deploytest.OkDeploymentConfig(1)
 	config.Spec.Strategy = recreateParams(30, "", deployapi.LifecycleHookFailurePolicyAbort, "")
 	deployment, _ := deployutil.MakeDeployment(config, kapi.Codecs.LegacyCodec(deployv1.SchemeGroupVersion))
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	hookExecuted := false
 	strategy := &RecreateDeploymentStrategy{
@@ -158,7 +158,7 @@ func TestRecreate_deploymentMidHookFail(t *testing.T) {
 	config := deploytest.OkDeploymentConfig(1)
 	config.Spec.Strategy = recreateParams(30, "", deployapi.LifecycleHookFailurePolicyAbort, "")
 	deployment, _ := deployutil.MakeDeployment(config, kapi.Codecs.LegacyCodec(deployv1.SchemeGroupVersion))
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	strategy := &RecreateDeploymentStrategy{
 		out:          &bytes.Buffer{},
@@ -190,7 +190,7 @@ func TestRecreate_deploymentPostHookSuccess(t *testing.T) {
 	config := deploytest.OkDeploymentConfig(1)
 	config.Spec.Strategy = recreateParams(30, "", "", deployapi.LifecycleHookFailurePolicyAbort)
 	deployment, _ := deployutil.MakeDeployment(config, kapi.Codecs.LegacyCodec(registered.GroupOrDie(kapi.GroupName).GroupVersions[0]))
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	hookExecuted := false
 	strategy := &RecreateDeploymentStrategy{
@@ -225,7 +225,7 @@ func TestRecreate_deploymentPostHookFail(t *testing.T) {
 	config := deploytest.OkDeploymentConfig(1)
 	config.Spec.Strategy = recreateParams(30, "", "", deployapi.LifecycleHookFailurePolicyAbort)
 	deployment, _ := deployutil.MakeDeployment(config, kapi.Codecs.LegacyCodec(registered.GroupOrDie(kapi.GroupName).GroupVersions[0]))
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	hookExecuted := false
 	strategy := &RecreateDeploymentStrategy{
@@ -258,7 +258,7 @@ func TestRecreate_deploymentPostHookFail(t *testing.T) {
 
 func TestRecreate_acceptorSuccess(t *testing.T) {
 	var deployment *kapi.ReplicationController
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	strategy := &RecreateDeploymentStrategy{
 		out:          &bytes.Buffer{},
@@ -304,7 +304,7 @@ func TestRecreate_acceptorSuccess(t *testing.T) {
 
 func TestRecreate_acceptorFail(t *testing.T) {
 	var deployment *kapi.ReplicationController
-	scaler := &scalertest.FakeScaler{}
+	scaler := &cmdtest.FakeScaler{}
 
 	strategy := &RecreateDeploymentStrategy{
 		out:          &bytes.Buffer{},
